requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.25.0
anthropic>=0.25.0

# Memory service dependencies
# Supabase support
supabase>=2.0.0

# Neo4j support
neo4j>=5.0.0

# Vector embeddings
sentence-transformers>=2.0.0
numpy>=1.24.0

# PDF processing
pymupdf>=1.24.0

# Optional: For better performance
# faiss-cpu>=1.7.0  # For faster similarity search
# torch>=2.0.0      # For sentence-transformers GPU support